NLP API - Stock Market News/stock_sentiment_nlp/requirements.txt
Flask==3.0.0
requests==2.31.0
nltk==3.8.1
finnhub-python==2.4.18
